create func.py and continue work in run_user_interface
